requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pillow
pytesseract
numpy